+++ algorithms/divide-and-conquer/max-subarray/msp-3.c
@@ -9,21 +9,21 @@
 
 #include "../../../lib/clib/clib.h"
 
-void printResult (int *arr, int l, int h, int max) {
+void printResult (int *arr, size_t l, size_t h, int max) {
   /* prints the result */
   printf("Maximum sum: %d\n", max);
   printf("Maximum subarray: \n[");
-  for (int i = l; i <= h; ++i) {
+  for (size_t i = l; i <= h; ++i) {
     printf("%d", arr[i]);
     if (i < h) printf(", ");
   }
   printf("]\n");
 }
 
-int maxSubarray(int *arr, int n, int *l, int *h) {
+int maxSubarray(int *arr, size_t n, size_t *l, size_t *h) {
   /* computes the maximum subarray */
   int max = INT_MIN, sum = 0, low = 0;
-  for (int i = 0; i < n; ++i) {
+  for (size_t i = 0; i < n; ++i) {
     sum += arr[i];
     if (sum > max) { // update max
       max = sum;
@@ -39,13 +39,13 @@ int maxSubarray(int *arr, int n, int *l, int *h) {
 }
 
 int main(int argc, char *argv[]) {
-  int n;
-  (void)! scanf("%d", &n);
+  size_t n;
+  (void)! scanf("%lu", &n);
 
   CREATE_ARRAY(int, arr, n);
   READ_ARRAY(arr, "%d", n);
 
-  int l = 0, h = n-1;  // low and high indices
+  size_t l = 0, h = n-1;  // low and high indices
   
   int max = maxSubarray(arr, n, &l, &h);
   printResult(arr, l, h, max);
